+++ pymdownx/util.py
@@ -101,16 +101,6 @@ else:
         return uchr(value)
 
 
-def get_arg_count(fn):
-    """Get argument count of function."""
-
-    if PY3:
-        count = fn.__code__.co_argcount
-    else:
-        count = fn.func_code.co_argcount
-    return count
-
-
 def escape_chars(md, echrs):
     """
     Add chars to the escape list.
@@ -186,5 +176,5 @@ def parse_url(url):
     return (scheme, netloc, path, params, query, fragment, is_url, is_absolute)
 
 
-class PymdownxDeprecationWarning(UserWarning):
+class PymdownxDeprecationWarning(UserWarning):  # pragma: no coverage
     """Deprecation warning for Pymdownx that is not hidden."""