+++ docs/src/markdown/changelog.md
@@ -3,6 +3,9 @@
 ## X.X.X
 
 - **NEW**: Details extension will now derive a title from the class if only a class is provided (https://github.com/facelessuser/pymdown-extensions/issues/107).
+- **NEW**: Remove deprecated legacy emoji generator format.
+- **NEW**: Remove deprecated `use_codehilite_settings`.
+- **NEW**: Remove deprecated `spoilers` extension redirect.
 
 ## 3.5.0